@@ -1,38 +1,37 @@
 {
-  "name": "huanui-notices-reader",
-  "version": "1.0.0",
-  "license": "MIT",
-  "scripts": {
-    "build": "rollup -c",
-    "clean": "rimraf public/*",
-    "dev": "yarn clean && rollup -c -w",
-    "dev:nowatch": "yarn build && yarn start",
-    "start": "sirv public --single",
-    "validate": "svelte-check"
-  },
-  "devDependencies": {
-    "@rollup/plugin-node-resolve": "^11.0.0",
-    "@rollup/plugin-typescript": "^8.0.0",
-    "@tsconfig/svelte": "^1.0.0",
-    "htmlnano": "^1.0.0",
-    "posthtml": "^0.16.5",
-    "posthtml-hash": "^1.2.2",
-    "rimraf": "^3.0.2",
-    "rollup": "^2.3.4",
-    "rollup-plugin-copy": "^3.4.0",
-    "rollup-plugin-css-only": "^3.1.0",
-    "rollup-plugin-livereload": "^2.0.0",
-    "rollup-plugin-serve": "^1.1.0",
-    "rollup-plugin-svelte": "^7.0.0",
-    "rollup-plugin-terser": "^7.0.0",
-    "svelte": "^3.0.0",
-    "svelte-calendar": "^2.0.4",
-    "svelte-check": "^1.0.0",
-    "svelte-preprocess": "^4.7.2",
-    "tslib": "^2.0.0",
-    "typescript": "^4.0.0"
-  },
-  "dependencies": {
-    "sass": "^1.43.5"
-  }
+	"name": "hcnoticereader",
+	"version": "0.0.1",
+	"scripts": {
+		"dev": "vite dev",
+		"build": "vite build",
+		"preview": "vite preview",
+		"test": "playwright test",
+		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
+		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
+		"lint": "eslint ."
+	},
+	"devDependencies": {
+		"@fontsource/fira-mono": "^4.5.10",
+		"@neoconfetti/svelte": "^1.0.0",
+		"@playwright/test": "^1.28.1",
+		"@sveltejs/adapter-auto": "^1.0.0",
+		"@sveltejs/adapter-cloudflare": "^1.0.0",
+		"@sveltejs/kit": "^1.0.0",
+		"@types/cookie": "^0.5.1",
+		"@typescript-eslint/eslint-plugin": "^5.45.0",
+		"@typescript-eslint/parser": "^5.45.0",
+		"dayjs": "^1.11.7",
+		"eslint": "^8.28.0",
+		"eslint-plugin-svelte3": "^4.0.0",
+		"sass": "^1.57.1",
+		"svelte": "^3.54.0",
+		"svelte-calendar": "^3.1.6",
+		"svelte-check": "^2.9.2",
+		"svelte-content-loader": "^1.1.3",
+		"svelte-simple-modal": "^1.4.5",
+		"tslib": "^2.4.1",
+		"typescript": "^4.9.3",
+		"vite": "^4.0.0"
+	},
+	"type": "module"
 }
